refactor: Organize relation related tests in one file
